// frontend/src/Types/ScholarTypes.tsx
export type UseCreateScholarType = {
    CreateScholarData: any;
    GetCreateScholarData: any;
}

export type UseGetMunicipalityType = {
    Municipalities: any;
    GetMunicipalities: any;
}

export type UseGetScholarType = {
    scholars: any;
    GetScholars: any;
}

export type UseBarangayType = {
    Barangays: any;
    GetBarangays: any;
}